both climb and user page scraping funcs functional
Ok, newest notebook just has the pure functions and they are working
well so far!

Check out how I set up the data storage, that can easily be changed...
Im not sure how we will want to do it when we try and make our
dataframes... in terms of conversion.

also, think about the different results and how we may or may not want
to standardize our data...